Fixed Body images so that they work

+++ MyRuns/src/edu/dartmouth/cs/myruns5/util/uv/ParseUVReading.java
@@ -16,8 +16,8 @@ public class ParseUVReading extends ParseObject{
 	public ParseUVReading() {
 	}
 
-	public int getUVI(){
-		return this.getInt(UVI);
+	public double getUVI(){
+		return this.getDouble(UVI);
 	}
 	
 	public ParseGeoPoint getLocation(){
